Cardano.Node.Emulator.Fee

Description

Calculating transaction fees in the emulator.

Documentation

estimateTransactionFee :: Params -> UTxO EmulatorEra -> [PaymentPubKeyHash] -> Tx -> Either CardanoLedgerError Value

estimateCardanoBuildTxFee :: Params -> UTxO EmulatorEra -> CardanoBuildTx -> Either CardanoLedgerError Value

makeAutoBalancedTransaction

Arguments

:: Params

-> UTxO EmulatorEra
Just the transaction inputs, not the entire UTxO.

-> CardanoBuildTx

-> CardanoAddress
Change address

-> Either CardanoLedgerError (Tx BabbageEra)

Creates a balanced transaction by calculating the execution units, the fees and the change, which is assigned to the given address. Only balances Ada.

makeAutoBalancedTransactionWithUtxoProvider

Arguments

:: Monad m

=> Params

-> UtxoIndex
Just the transaction inputs, not the entire UTxO.

-> CardanoAddress
Change address

-> (Value -> m ([(TxOutRef, TxOut)], Value))
The utxo provider, it return outputs that cover at least the given value, and return the change, i.e. how much the outputs overshoot the given value.

-> (forall a. CardanoLedgerError -> m a)
How to handle errors

-> CardanoBuildTx

-> m (Tx BabbageEra)

Creates a balanced transaction by calculating the execution units, the fees and then the balance. If the balance is negative the utxo provider is asked to pick extra inputs to make the balance is positive, which is then assigned to the change address. The collateral is similarly balanced. Unlike makeAutoBalancedTransaction this function also balances non-Ada.

utxoProviderFromWalletOutputs

Arguments

:: Map TxOutRef TxOut
The unspent transaction outputs. Make sure that this doesn't contain any inputs from the transaction being balanced.

-> Value

-> Either BalancingError ([(TxOutRef, TxOut)], Value)

data BalancingError

Constructors

InsufficientFunds
Not enough extra inputs available to balance a transaction.

Fields

total :: Value

expected :: Value

CardanoLedgerError CardanoLedgerError

Internals

selectCoin

Arguments

:: Eq a

=> [(a, Value)]
Possible inputs to choose from

-> Value
The target value

-> Either BalancingError ([a], Value)
The chosen inputs and the change

Given a set of as with coin values, and a target value, select a number of a such that their total value is greater than or equal to the target.