Cardano.Node.Emulator.Internal.API

Description

If you want to run the node emulator without using the Contract monad, this module provides a simple MTL-based interface.

Types

data EmulatorState

Constructors

EmulatorState !ChainState !AddressMap !(Map DatumHash Datum)

esChainState :: Lens' EmulatorState ChainState

esAddressMap :: Lens' EmulatorState AddressMap

esDatumMap :: Lens' EmulatorState (Map DatumHash Datum)

data EmulatorError

Constructors

BalancingError !BalancingError

ValidationError !ValidationErrorInPhase

ToCardanoError !ToCardanoError

type EmulatorLogs = Seq (LogMessage EmulatorMsg)

data EmulatorMsg

Constructors

GenericMsg Value

ChainEvent ChainEvent

data LogMessage a

Constructors

LogMessage

Fields

_logLevel :: LogLevel

_logMessageContent :: a

type MonadEmulator m = (MonadRWS Params EmulatorLogs EmulatorState m, MonadError EmulatorError m)

type EmulatorT m = ExceptT EmulatorError (RWST Params EmulatorLogs EmulatorState m)

type EmulatorM = EmulatorT Identity

Running Eff chain effects in MTL

handleChain :: MonadEmulator m => Eff [ChainControlEffect, ChainEffect] a -> m a

processBlock :: MonadEmulator m => m Block

modifySlot :: MonadEmulator m => (Slot -> Slot) -> m Slot