Cardano.Node.Emulator.Internal.Node.TimeSlot

Documentation

data SlotConfig

Datatype to configure the length (ms) of one slot and the beginning of the first slot.

Constructors

SlotConfig

Fields

scSlotLength :: !Integer

Length (number of milliseconds) of one slot

scSlotZeroTime :: !POSIXTime

Beginning of slot 0 (in milliseconds)

data SlotConversionError

Constructors

SlotOutOfRange

Fields

requestedSlot :: !Slot

horizon :: !(Slot, POSIXTime)

slotRangeToPOSIXTimeRange :: SlotConfig -> SlotRange -> POSIXTimeRange

Convert a SlotRange to a POSIXTimeRange given a SlotConfig. The resulting POSIXTimeRange refers to the starting time of the lower bound of the SlotRange and the ending time of the upper bound of the SlotRange.

slotToPOSIXTimeRange :: SlotConfig -> Slot -> POSIXTimeRange

Convert a Slot to a POSIXTimeRange given a SlotConfig. Each Slot can be represented by an interval of time.

slotToBeginPOSIXTime :: SlotConfig -> Slot -> POSIXTime

Get the starting POSIXTime of a Slot given a SlotConfig.

slotToEndPOSIXTime :: SlotConfig -> Slot -> POSIXTime

Get the ending POSIXTime of a Slot given a SlotConfig.

posixTimeRangeToContainedSlotRange :: SlotConfig -> POSIXTimeRange -> SlotRange

Convert a POSIXTimeRange to SlotRange given a SlotConfig. This gives the biggest slot range that is entirely contained by the given time range.

posixTimeToEnclosingSlot :: SlotConfig -> POSIXTime -> Slot

Convert a POSIXTime to Slot given a SlotConfig.

currentSlot :: SlotConfig -> IO Slot

Get the current slot number

utcTimeToPOSIXTime :: UTCTime -> POSIXTime

posixTimeToUTCTime :: POSIXTime -> UTCTime

nominalDiffTimeToPOSIXTime :: NominalDiffTime -> POSIXTime

posixTimeToNominalDiffTime :: POSIXTime -> NominalDiffTime