Cardano.Node.Emulator.Internal.Node.Validation

Description

Transaction validation using 'cardano-ledger-specs'

Documentation

type EmulatorBlock = [Validated (Tx EmulatorEra)]

data EmulatedLedgerState

State of the ledger with configuration, mempool, and the blockchain.

Constructors

EmulatedLedgerState

Fields

_ledgerEnv :: !(MempoolEnv EmulatorEra)

_memPoolState :: !(MempoolState EmulatorEra)

_currentBlock :: !EmulatorBlock

_previousBlocks :: ![EmulatorBlock]

newtype Coin

Constructors

Coin

Fields

unCoin :: Integer

newtype SlotNo

Constructors

SlotNo

Fields

unSlotNo :: Word64

type EmulatorEra = BabbageEra StandardCrypto

The default era for the emulator

type CardanoLedgerError = Either ValidationErrorInPhase ToCardanoError

initialState :: Params -> EmulatedLedgerState

Initial ledger state for a distribution

hasValidationErrors :: Params -> SlotNo -> UtxoIndex -> Tx BabbageEra -> ValidationResult

makeTransactionBody :: Params -> UTxO EmulatorEra -> CardanoBuildTx -> Either CardanoLedgerError (TxBody BabbageEra)

validateCardanoTx :: Params -> Slot -> UtxoIndex -> CardanoTx -> ValidationResult

unsafeMakeValid :: CardanoTx -> OnChainTx

Modifying the state

makeBlock :: EmulatedLedgerState -> EmulatedLedgerState

Make a block with all transactions that have been validated in the current block, add the block to the blockchain, and empty the current block.

setSlot :: SlotNo -> EmulatedLedgerState -> EmulatedLedgerState

Set the slot number

nextSlot :: EmulatedLedgerState -> EmulatedLedgerState

Increase the slot number by one

newtype UTxO era

Constructors

UTxO

Fields

unUTxO :: Map (TxIn (Crypto era)) (TxOut era)

setUtxo :: UTxO EmulatorEra -> EmulatedLedgerState -> EmulatedLedgerState

Set the utxo

Lenses

ledgerEnv :: Lens' EmulatedLedgerState (MempoolEnv EmulatorEra)

memPoolState :: Lens' EmulatedLedgerState (MempoolState EmulatorEra)

currentBlock :: Lens' EmulatedLedgerState EmulatorBlock

previousBlocks :: Lens' EmulatedLedgerState [EmulatorBlock]

Etc.

emulatorGlobals :: Params -> Globals

A sensible default Globals value for the emulator