Cardano.Node.Emulator.MTL

Description

If you want to run the node emulator without using the Contract monad, this module provides a simple MTL-based interface.

Updating the blockchain

queueTx :: MonadEmulator m => CardanoTx -> m ()

Queue the transaction, it will be processed when nextSlot is called.

nextSlot :: MonadEmulator m => m ()

Process the queued transactions and increase the slot number.

currentSlot :: MonadEmulator m => m Slot

Get the current slot number of the emulated node.

awaitSlot :: MonadEmulator m => Slot -> m ()

Call nextSlot until the current slot number equals or exceeds the given slot number.

Querying the blockchain

utxosAt :: MonadEmulator m => CardanoAddress -> m (Map TxOutRef DecoratedTxOut)

Query the unspent transaction outputs at the given address.

fundsAt :: MonadEmulator m => CardanoAddress -> m Value

Query the total value of the unspent transaction outputs at the given address.

Transactions

balanceTx

Arguments

:: MonadEmulator m

=> UtxoIndex
Just the transaction inputs, not the entire UTxO.

-> CardanoAddress
Wallet address

-> CardanoBuildTx

-> m CardanoTx

Balance an unbalanced transaction, using funds from the given wallet if needed, and returning any remaining value to the same wallet.

signTx

Arguments

:: (MonadEmulator m, Foldable f)

=> f PaymentPrivateKey
Signatures

-> CardanoTx

-> m CardanoTx

Sign a transaction with the given signatures.

submitUnbalancedTx

Arguments

:: (MonadEmulator m, Foldable f)

=> UtxoIndex
Just the transaction inputs, not the entire UTxO.

-> CardanoAddress
Wallet address

-> f PaymentPrivateKey
Signatures

-> CardanoBuildTx

-> m CardanoTx

Balance a transaction, sign it with the given signatures, and finally queue it.

submitTxConfirmed

Arguments

:: (MonadEmulator m, Foldable f)

=> UtxoIndex
Just the transaction inputs, not the entire UTxO.

-> CardanoAddress

-> f PaymentPrivateKey

-> CardanoBuildTx

-> m CardanoTx

payToAddress :: MonadEmulator m => (CardanoAddress, PaymentPrivateKey) -> CardanoAddress -> Value -> m TxId

Create a transaction that transfers funds from one address to another, and sign and submit it.

Logging

logDebug :: (ToJSON a, MonadEmulator m) => a -> m ()

Log a message at the Debug level

logInfo :: (ToJSON a, MonadEmulator m) => a -> m ()

Log a message at the Info level

logWarn :: (ToJSON a, MonadEmulator m) => a -> m ()

Log a message at the Warning level

logError :: (ToJSON a, MonadEmulator m) => a -> m ()

Log a message at the Error level

Types

data EmulatorState

Constructors

EmulatorState !ChainState !AddressMap

esChainState :: Lens' EmulatorState ChainState

esAddressMap :: Lens' EmulatorState AddressMap

data EmulatorError

Constructors

BalancingError !BalancingError

ValidationError !ValidationErrorInPhase

ToCardanoError !ToCardanoError

type EmulatorLogs = Seq (LogMessage EmulatorMsg)

data EmulatorMsg

Constructors

GenericMsg Value

TxBalanceMsg TxBalanceMsg

ChainEvent ChainEvent

data LogMessage a

Constructors

LogMessage

Fields

_logLevel :: LogLevel

_logMessageContent :: a

type MonadEmulator m = (MonadRWS Params EmulatorLogs EmulatorState m, MonadError EmulatorError m)

type EmulatorT m = ExceptT EmulatorError (RWST Params EmulatorLogs EmulatorState m)

type EmulatorM = EmulatorT Identity

emptyEmulatorState :: EmulatorState

emptyEmulatorStateWithInitialDist :: Map CardanoAddress Value -> EmulatorState

getParams :: MonadEmulator m => m Params

Running Eff chain effects in MTL

handleChain :: MonadEmulator m => Eff [ChainControlEffect, ChainEffect] a -> m a