Cardano.Node.Emulator.MTL.Test

Basic testing

hasValidatedTransactionCountOfTotal :: Int -> Int -> EmulatorLogs -> Maybe String

Test the number of validated transactions and the total number of transactions. Returns a failure message if the numbers don't match up.

renderLogs :: EmulatorLogs -> Text

Render the logs in a format useful for debugging why a test failed.

Testing with `quickcheck-contractmodel`

propSanityCheckModel :: forall state. ContractModel state => Property

Sanity check a ContractModel. Ensures that wallet balances are not always unchanged.

propSanityCheckAssertions :: forall state. ContractModel state => Actions state -> Property

Sanity check a ContractModel. Ensures that all assertions in the property generation succeed.

propRunActions_

Arguments

:: forall state. RunModel state EmulatorM

=> Actions state
The actions to run

-> Property

Run Actions in the emulator and check that the model and the emulator agree on the final wallet balance changes. Starts with 100.000.000 Ada for each wallet and the default parameters.

propRunActions

Arguments

:: forall state. RunModel state EmulatorM

=> (ModelState state -> EmulatorLogs -> Maybe String)
Predicate to check at the end of execution

-> Actions state
The actions to run

-> Property

propRunActionsWithOptions

Arguments

:: forall state. RunModel state EmulatorM

=> Map CardanoAddress Value
Initial distribution of funds

-> Params
Node parameters

-> (ModelState state -> EmulatorLogs -> Maybe String)
Predicate to check at the end of execution

-> Actions state
The actions to run

-> Property

Other exports

chainStateToChainIndex :: NetworkId -> ChainState -> ChainIndex

chainStateToContractModelChainState :: ChainState -> ChainState

Orphan instances

IsRunnable EmulatorM

Instance details

Methods

awaitSlot :: SlotNo -> EmulatorM ()

HasChainIndex EmulatorM

Instance details

Methods

getChainIndex :: EmulatorM ChainIndex

getChainState :: EmulatorM ChainState