Cardano.Node.Emulator.Params

Description

The set of parameters, like protocol parameters and slot configuration.

Documentation

data Params

Constructors

Params

Fields

pSlotConfig :: SlotConfig

emulatorPParams :: PParams

Convert Params to cardano-ledger PParams

pNetworkId :: NetworkId

paramsWithProtocolsParameters :: SlotConfig -> ProtocolParameters -> NetworkId -> Params

slotConfigL :: Lens' Params SlotConfig

emulatorPParamsL :: Lens' Params PParams

pParamsFromProtocolParams :: ProtocolParameters -> PParams

pProtocolParams :: Params -> ProtocolParameters

protocolParamsL :: Lens' Params ProtocolParameters

networkIdL :: Lens' Params NetworkId

increaseTransactionLimits :: Params -> Params

Set higher limits on transaction size and execution units. This can be used to work around MaxTxSizeUTxO and ExUnitsTooBigUTxO errors. Note that if you need this your Plutus script will probably not validate on Mainnet.

genesisDefaultsFromParams :: Params -> ShelleyGenesis EmulatorEra

cardano-ledger specific types and conversion functions

type EmulatorEra = BabbageEra StandardCrypto

The default era for the emulator

type PParams = PParams EmulatorEra

slotLength :: Params -> SlotLength

Calculate the cardano-ledger SlotLength

testnet :: NetworkId

The network id used by default by Param

emulatorEpochSize :: EpochSize

A sensible default EpochSize value for the emulator

emulatorGlobals :: Params -> Globals

A sensible default Globals value for the emulator

emulatorEraHistory :: Params -> EraHistory CardanoMode

A sensible default EraHistory value for the emulator

Orphan instances

ToJSON NetworkId

Instance details

Methods

toJSON :: NetworkId -> Value

toEncoding :: NetworkId -> Encoding

toJSONList :: [NetworkId] -> Value

toEncodingList :: [NetworkId] -> Encoding

ToJSON NetworkMagic

Instance details

Methods

toJSON :: NetworkMagic -> Value

toEncoding :: NetworkMagic -> Encoding

toJSONList :: [NetworkMagic] -> Value

toEncodingList :: [NetworkMagic] -> Encoding

FromJSON NetworkId

Instance details

Methods

parseJSON :: Value -> Parser NetworkId

parseJSONList :: Value -> Parser [NetworkId]

FromJSON NetworkMagic

Instance details

Methods

parseJSON :: Value -> Parser NetworkMagic

parseJSONList :: Value -> Parser [NetworkMagic]

Default ProtocolParameters

Instance details

Methods

def :: ProtocolParameters