Cardano.Node.Emulator.Validation

Description

Transaction validation using 'cardano-ledger-specs'

Documentation

type EmulatorBlock = [Validated (Tx EmulatorEra)]

data EmulatedLedgerState

State of the ledger with configuration, mempool, and the blockchain.

Constructors

EmulatedLedgerState

Fields

_ledgerEnv :: MempoolEnv EmulatorEra

_memPoolState :: MempoolState EmulatorEra

_currentBlock :: EmulatorBlock

_previousBlocks :: [EmulatorBlock]

newtype Coin

The amount of value held by a transaction output.

Constructors

Coin

Fields

unCoin :: Integer

newtype SlotNo

The 0-based index for the Ourboros time slot.

Constructors

SlotNo

Fields

unSlotNo :: Word64

type EmulatorEra = BabbageEra StandardCrypto

The default era for the emulator

type CardanoLedgerError = Either ValidationErrorInPhase ToCardanoError

initialState :: Params -> EmulatedLedgerState

Initial ledger state for a distribution

hasValidationErrors :: Params -> SlotNo -> UTxO EmulatorEra -> Tx BabbageEra -> Either ValidationErrorInPhase ValidationSuccess

makeTransactionBody :: Params -> UTxO EmulatorEra -> CardanoBuildTx -> Either CardanoLedgerError (TxBody BabbageEra)

validateCardanoTx :: Params -> Slot -> UTxO EmulatorEra -> CardanoTx -> Either ValidationErrorInPhase ValidationSuccess

fromPlutusTx :: Params -> UTxO EmulatorEra -> [PaymentPubKeyHash] -> Tx -> Either CardanoLedgerError (Tx BabbageEra)

fromPlutusTxSigned :: Params -> UTxO EmulatorEra -> Tx -> Map PaymentPubKey PaymentPrivateKey -> CardanoTx

fromPlutusTxSigned' :: Params -> UTxO EmulatorEra -> Tx -> Map PaymentPubKey PaymentPrivateKey -> Either CardanoLedgerError CardanoTx

Modifying the state

makeBlock :: EmulatedLedgerState -> EmulatedLedgerState

Make a block with all transactions that have been validated in the current block, add the block to the blockchain, and empty the current block.

setSlot :: SlotNo -> EmulatedLedgerState -> EmulatedLedgerState

Set the slot number

nextSlot :: EmulatedLedgerState -> EmulatedLedgerState

Increase the slot number by one

newtype UTxO era

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO :: Map (TxIn (Crypto era)) (TxOut era)

setUtxo :: UTxO EmulatorEra -> EmulatedLedgerState -> EmulatedLedgerState

Set the utxo

Lenses

ledgerEnv :: Lens' EmulatedLedgerState (MempoolEnv EmulatorEra)

memPoolState :: Lens' EmulatedLedgerState (MempoolState EmulatorEra)

currentBlock :: Lens' EmulatedLedgerState EmulatorBlock

previousBlocks :: Lens' EmulatedLedgerState [EmulatorBlock]

Etc.

emulatorGlobals :: Params -> Globals

A sensible default Globals value for the emulator