Cardano.Node.Emulator.Chain

Documentation

data ChainEvent

Events produced by the blockchain emulator.

Constructors

TxnValidate TxId CardanoTx [Text]
A transaction has been validated and added to the blockchain.

TxnValidationFail ValidationPhase TxId CardanoTx ValidationError Value [Text]
A transaction failed to validate. The Value indicates the amount of collateral stored in the transaction.

SlotAdd Slot

type TxPool = [CardanoTx]

A pool of transactions which have yet to be validated.

data ChainState

Constructors

ChainState

Fields

_chainNewestFirst :: Blockchain

The current chain, with the newest transactions first in the list.

_txPool :: TxPool

The pool of pending transactions.

_index :: UtxoIndex

The UTxO index, used for validation.

_chainCurrentSlot :: Slot

The current slot number

emptyChainState :: ChainState

txPool :: Lens' ChainState TxPool

index :: Lens' ChainState UtxoIndex

chainNewestFirst :: Lens' ChainState Blockchain

chainCurrentSlot :: Lens' ChainState Slot

data ChainControlEffect r where

Constructors

ProcessBlock :: ChainControlEffect Block

ModifySlot :: (Slot -> Slot) -> ChainControlEffect Slot

data ChainEffect r where

Constructors

QueueTx :: CardanoTx -> ChainEffect ()

GetCurrentSlot :: ChainEffect Slot

GetParams :: ChainEffect Params

processBlock :: Member ChainControlEffect effs => Eff effs Block

Make a new block

modifySlot :: Member ChainControlEffect effs => (Slot -> Slot) -> Eff effs Slot

Adjust the current slot number, returning the new slot.

queueTx :: Member ChainEffect effs => CardanoTx -> Eff effs ()

getParams :: Member ChainEffect effs => Eff effs Params

getCurrentSlot :: Member ChainEffect effs => Eff effs Slot

type ChainEffs = '[State ChainState, LogMsg ChainEvent]

handleControlChain :: Members ChainEffs effs => Params -> ChainControlEffect ~> Eff effs

logEvent :: Member (LogMsg ChainEvent) effs => ChainEvent -> Eff effs ()

handleChain :: Members ChainEffs effs => Params -> ChainEffect ~> Eff effs

data ValidatedBlock

The result of validating a block.

Constructors

ValidatedBlock

Fields

vlbValid :: Block

The transactions that have been validated in this block.

vlbEvents :: [ChainEvent]

Transaction validation events for the transactions in this block.

vlbIndex :: UtxoIndex

The updated UTxO index after processing the block

data ValidationCtx

Constructors

ValidationCtx

Fields

vctxIndex :: UtxoIndex

vctxParams :: Params

validateBlock :: Params -> Slot -> UtxoIndex -> TxPool -> ValidatedBlock

Validate a block given the current slot and UTxO index, returning the valid transactions, success/failure events and the updated UTxO set.

getCollateral :: UtxoIndex -> CardanoTx -> Value

canValidateNow :: Slot -> CardanoTx -> Bool

Check whether the given transaction can be validated in the given slot.

mkValidationEvent :: UtxoIndex -> CardanoTx -> Either ValidationErrorInPhase ValidationSuccess -> ChainEvent

validateEm :: MonadState ValidationCtx m => Slot -> CardanoTx -> m (Either ValidationErrorInPhase ValidationSuccess)

Validate a transaction in the current emulator state.

addBlock :: Block -> ChainState -> ChainState

Adds a block to ChainState, without validation.

addTxToPool :: CardanoTx -> TxPool -> TxPool

_SlotAdd :: Prism' ChainEvent Slot

_TxnValidationFail :: Prism' ChainEvent (ValidationPhase, TxId, CardanoTx, ValidationError, Value, [Text])

_TxnValidate :: Prism' ChainEvent (TxId, CardanoTx, [Text])